Added renderHistorySections as a part of utils/TimeUtils.jsx => changed from utils/FirebaseUtils.jsx
